add double for rating in migrations

+++ database/migrations/2014_10_12_000006_create_users_table.php
@@ -24,7 +24,7 @@ class CreateUsersTable extends Migration
             $table->bigInteger('lang_id')->unsigned()->index()->nullable();
             $table->foreign('lang_id')->references('id')->on('lang');
             $table->integer('order')->default(1);
-            $table->integer('rating')->default(0);
+            $table->double('rating')->default(0);
             $table->integer('rating_values')->default(0);
             $table->integer('rating_votes')->default(0);
             $table->boolean('is_changed')->default(false);
